Implement handleSubmit function
