+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/aldana/Microverse/bookstore/src/components/BookForm.js":"1"},{"size":1004,"mtime":1611648168576,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":1,"warningCount":0,"fixableErrorCount":1,"fixableWarningCount":0,"source":null},"ipiu3q","/home/aldana/Microverse/bookstore/src/components/BookForm.js",["6"],{"ruleId":"7","severity":2,"message":"8","line":23,"column":82,"nodeType":"9","fix":"10"},"react/jsx-tag-spacing","A space is required before closing bracket","JSXOpeningElement",{"range":"11","text":"12"},[578,578]," "]
+[{"/home/aldana/Microverse/bookstore/src/containers/BookForm.js":"1","/home/aldana/Microverse/bookstore/src/index.js":"2","/home/aldana/Microverse/bookstore/src/components/Book.js":"3","/home/aldana/Microverse/bookstore/src/redux/actions/index.js":"4","/home/aldana/Microverse/bookstore/src/components/App.js":"5","/home/aldana/Microverse/bookstore/src/containers/BooksList.js":"6"},{"size":1653,"mtime":1611651076055,"results":"7","hashOfConfig":"8"},{"size":360,"mtime":1611650729327,"results":"9","hashOfConfig":"8"},{"size":455,"mtime":1611587890692,"results":"10","hashOfConfig":"8"},{"size":233,"mtime":1611587890692,"results":"11","hashOfConfig":"8"},{"size":257,"mtime":1611650755767,"results":"12","hashOfConfig":"8"},{"size":699,"mtime":1611651021471,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"ipiu3q",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/home/aldana/Microverse/bookstore/src/containers/BookForm.js",["26"],"/home/aldana/Microverse/bookstore/src/index.js",[],"/home/aldana/Microverse/bookstore/src/components/Book.js",[],"/home/aldana/Microverse/bookstore/src/redux/actions/index.js",[],"/home/aldana/Microverse/bookstore/src/components/App.js",[],"/home/aldana/Microverse/bookstore/src/containers/BooksList.js",["27"],{"ruleId":"28","severity":1,"message":"29","line":30,"column":5,"nodeType":"30","messageId":"31","endLine":30,"endColumn":16},{"ruleId":"28","severity":1,"message":"29","line":32,"column":3,"nodeType":"30","messageId":"31","endLine":32,"endColumn":14},"no-console","Unexpected console statement.","MemberExpression","unexpected"]
